requirements.txt
absl-py==1.0.0
albumentations==1.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.2.0
backcall==0.2.0
black==21.12b0
bleach==4.1.0
blobfile==1.2.7
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
contextlib2==21.6.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
efficientnet-pytorch==0.6.3
einops==0.3.2
entrypoints==0.3
facenet-pytorch==2.5.2
filelock==3.4.0
fonttools==4.29.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
huggingface-hub==0.4.0
idna==3.3
imageio==2.13.5
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutils==0.5.4
iniconfig==1.1.1
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.3
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
ml-collections==0.1.1
munch==2.5.0
mypy==0.920
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.7
notebook==6.4.6
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626271506491/work
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.4.60
opencv-python-headless==4.5.5.62
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
pretrainedmodels==0.7.4
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.12.0
pydicom==2.3.0
Pygments==2.11.1
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
qudida==0.0.4
regex==2022.3.15
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.49
scikit-image==0.19.2
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
sentencepiece==0.1.96
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
submitit==1.4.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2022.2.9
timm==0.3.2
tokenizers==0.11.5
toml==0.10.2
tomli==1.2.3
torch==1.8.2
torchaudio==0.8.2
torchvision==0.9.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.17.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
xmltodict==0.12.0
zipp==3.6.0
